Prediction_API/requirements.txt
asgiref==3.5.2
click==8.1.3
colorama==0.4.5
Django==3.2.14
djangorestframework==3.13.1
importlib-metadata==4.12.0
joblib==1.1.0
nltk==3.7
numpy==1.21.6
pandas==1.3.5
python-dateutil==2.8.2
pytz==2022.1
regex==2022.7.9
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tqdm==4.64.0
typing-extensions==4.3.0
zipp==3.8.0
